Added Controller Handlers
Applications now have an app/handlers/ directory, containing route handlers for
each controller. Handlers are route callbacks stored in separate files, which
serve as an organization scheme for controllers with huge amounts of code.

Handlers can be referenced when defining routes:

  get('/test', this.handler('test.js'));

Additionally, a path relative to the controller's handler directory can be passed:

  get('/test', this.handler('misc/directory/handler.js'));

Use the 'protos handler' command to create handlers:

  $ protos handler main/hello,world,there admin/some/path/to/handler.js

Each handler module exports the callback function which itself is the route function
to be called when a given route matches.