Implemented raw routes
Raw routes are prefixed with `raw_`, and are special, compared to the
other route functions: They are run just after the route has been matched,
and no extra or automatic processing from the controller is done.

The following operations are not performed by the controller for raw routes:

- No request body processing (if body_parser middleware is loaded)
- No filters executed
- No session loaded automatically (if session middleware is enabled)
- Authentication is ignored

Basically, raw routes are used in case there is a need to do things manually
for a specific route, without using the framework's internal processing.

As soon as a route matches, raw routes are executed, and any controller processing
is ignored on the spot.

The route functions for raw routes are the same as the other route functions
(matching the HTTP methods in lowercase), but this time prefixed with `raw_`,
for example: raw_get, raw_post, etc.