Added req.noSessionRegenerate()
The method should be run before loading the session (that is,
before `app.session.loadSession` or the route callback `app.session.load`.

When called, the session will not regenerate the session id if the session
hash cookie is not present on the request.

This is of extreme importance when dealing with concurrent operations to
a specific controller. If the session hash expires during one of the
requests, the session is regenerated, which will invalidate any of the
pending request that may hit the server after the session has been regenerated
since they will come with a recently expired session ID.

If the route prevents the session from being regenerated whenever these calls
(usually concurrent AJAX calls) reach it, there won't be any conflicts with
session regeneration.