Make sure empty validator msg never returns falsy
This ensures that at least an invalid message is outputted. It is
a convention that the validator returns null on success. If a validation
message returns undefined, then it would be assumed that the validation
has succeeded, when it has not.

Forcing the validation method to return an invalid message is the right
thing to do, to prevent invalid messages to be confused with valid ones.