Implemented Cache Filters
Cache filters are callbacks that modify the query results before
caching them in the underlying storage. This results in increased
performance, since the filter only runs before caching the callback,
instead of running each time the cache is retrieved.

To specify a Cache Filter, add it to the Cache Data object passed
to the Driver::queryCached callback, as such:

db.queryCached({
  cacheID: 'my_cache',
  cacheFilter: function(data) {
    // Modify the data object, or return a new object
    return data;
  }
}, 'exec', function(err, data) {
  // The data object is already modified by the filter
  console.log([err, data]);
});