Relocated skeleton/lib to skeleton/app/properties
This change has been motivated due to the necessity to make
use of standalone libraries not available as npm modules.

The lib/ directory turns out to be a bit misleading when it
comes to the functionality given initially. It has now been
repurposed to store standalone .js files or libraries.

The 'protos lib' command has been removed to avoid confusion.