Added 'worker_message' and 'master_message' events
These events are emitted when running on a clustered environment.

When a message from a worker arrives at the master process, the
'worker_message' event is emitted on the master process, with
[msg, worker] as parameters.

When a message from the master process arrives at a worker process,
the 'master_message' event is emitted on the worker process, with
[msg] as parameter.

When running on a clustered environment, there is a back and forth
communication between the master process and its workers, using
the two events internally to communicate.

This serves the additional purpose of testing the communication
channels between master & workers. If the communication is not
working, the worker servers will not be started.

Additionally, debugging messages are printed to the stdout, when
the internal events (used to start the worker servers) are unbound.