Application::invalidateCacheInterval now uses transactions
Before, a multi object was used to delete each cacheID passed. This doesn't
use transactions and a new connection to the database needs to be established
on each operation.

Now, the invalidateCacheInterval method uses the internal storage's transaction
mechanisms to delete multiple cache entries, which results in better performance
and a single connection being established to perform multiple delete operations.