Add env command to protos command line tool
Environment data can be retrieved from the command line by using
the protos env command, similarly to how it is used in javascript:

$ protos env postgres.username

Will retrieve the username key within the postgres object on the
environment data file, which by default is development.json

To specify a custom JSON file, pass the src option:

$ protos env postgres.username --src myconfig.json

The command runs anywhere from the application's path and the
paths provided to the src option are relative to the application's
root directory.

The protos env command can be used in shell scripts and Makefiles
to retrieve environment variables from the application.