Add asset helper via asset compiler middleware
The asset helper can be used in views as such:

{{asset src="/core/css/test.css"}}

This will output something like this:

/core/css/test-b1946ac92492d2347c6235b4d2611184.css

This is useful when running on production, to properly
handle caching of files based on their modification time.

The suffix is calculated using the md5 hash of the file's
modification time (without milliseconds).

The mstat hashes are cached, the cache can be cleared by
using hot code reload, as such:

app.reload({assets: true});

This will cause the hashes to be updated if the modification
times of the files have changed.