The postSuffix configuration is avaliable on the server section of the boot.js file, and is used to remove the port suffix from the application's URL.
The application's baseUrl and baseUrlSecure options will have the
port included in the URL if the option is set to true (default).

In addition, URLs returned by app.url() will also have the port
suffix removed from the URLs.

For backwards compatibility with previous protos versions, if the
portSuffix option is not present on the boot.js file, it will
be assumed as enabled (the same is valid if it's set to undefined).

The default behavior has always been to display the port suffix
in the URLs. The portSuffix option provides greater control for
the specific situations in which the port suffix must not be
included in URLs, when running behind proxies for example.