Add config.transientRaw option to render transient views
This option will provide more control to render transient views,
which are 404, 500 and regular messages. The application methods
that render transient views are:

- app.notFound();
- app.serverError();
- res.httpMessage();

Prior to this commit, such messages were always rendered in raw,
which means the header/footer templates were never included.

The config.transientRaw option controls whether or not to display
the transient views in raw.

If config.transientRaw is false, it will automatically take the
value of config.rawViews, to adapt properly for global raw views.

Additionally, the transient_view event is emitted each time a
transient view is about to be rendered, which is useful to perform
specific actions prior to the views being rendered.